12 VersionParsingBug
On some systems the version of Python is just two numbers, e.g 2.7 and
not three numbers as normally expected, e.g 2.7.16. This caused the
version parser to try to parse an integer from unexpected input.

The new version detects a space between the numbers and this should fix
the issue.