Added a custom debug server for "Debug Adapter Protocol"
Added a custom debugging server to replace pydebug. This has the advantage that we can control which threads are being enablet by settrace and we dont have to call extremely slow python code whenever we think we might need to call settrace.

This fixes the issue we had related to debugging not working before calling 'debug_this_thread.'